secrets_backend/scanning/requirements.txt
GitPython
requests
pymongo